implement accessryitem delete

+++ src/main/java/demo/bfims/Controllers/AccessoryController.java
@@ -3,7 +3,6 @@ package demo.bfims.Controllers;
 import demo.bfims.Config.Response;
 import demo.bfims.DTOs.InventoryDTOs.Accessory.AccessoryDto;
 import demo.bfims.DTOs.InventoryDTOs.Accessory.AccessoryItemDto;
-import demo.bfims.Repo.AccessoryRepo;
 import demo.bfims.Services.AccessoryService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
@@ -16,29 +15,50 @@ public class AccessoryController {
     @Autowired
     AccessoryService accessoryService;
 
+    //Get list of all accessories
     @GetMapping("/accessories")
+    public List<AccessoryDto> getAccessories() {
+        return accessoryService.getAccessories();
+    }
+
+    //Delete accessory
+    @DeleteMapping("/accessory/{id}")
+    public Response deleteAccessory(@PathVariable Long id) {
+        if(id != null) return accessoryService.removeAccessory(id);
+        else return null;
+    }
+
+    // get specific accessory
+    @GetMapping("/accessory/{id}")
+    public AccessoryDto newAccessory(@PathVariable Long id) {
+        if(id != null) return accessoryService.getAccessory(id);
+        else return null;
+    }
+
+    //Update Accessory
+    @PutMapping("/accessory")
+    public AccessoryDto updateAccessory(@RequestBody AccessoryDto accessoryDto) {
+        if (accessoryDto != null) return accessoryService.updateAccessory(accessoryDto);
+        else return null;
+    }
+
+    //Get all accessoryItems (may be useless)
+    @GetMapping("/accessoryitems")
     public List<AccessoryItemDto> getAccessoryItems() {
         return accessoryService.getAccessoryItems();
     }
 
-    @PostMapping("/accessory")
+    //Create new accessoryItem. Creates new accessory by default
+    @PostMapping("/accessoryitem")
     public AccessoryItemDto newAccessory(@RequestBody AccessoryItemDto accessoryItemDto) {
-       return accessoryService.newAccessory(accessoryItemDto);
+        if(accessoryItemDto != null) return accessoryService.newAccessory(accessoryItemDto);
+        else return null;
     }
 
-    @DeleteMapping("/accessory/{id}")
+    @DeleteMapping("/accessoryitem/{id}")
     public Response deleteAccessoryItem(@PathVariable Long id) {
-        return accessoryService.removeAccessory(id);
+        if(id != null) return accessoryService.removeAccessoryItem(id);
+        else return null;
     }
 
-    @GetMapping("/accessory/{id}")
-    public AccessoryDto newAccessory(@PathVariable Long id) {
-       return accessoryService.getAccessory(id);
-    }
-
-    @PutMapping("/accessory")
-    public AccessoryDto updateAccessory(@RequestBody AccessoryDto accessoryDto) {
-        return accessoryService.updateAccessory(accessoryDto);
-
-    }
 }
